depthModels/pyproject.toml
[tool.poetry]
name = "depthmodels"
version = "0.1.0"
description = ""
authors = ["Herrick Fang <herrickfang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
scipy = "^1.12.0"
opencv-python = "^4.9.0.80"
torch = "^2.2.0"
moviepy = "^1.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
